requirements.txt
blinker==1.7.0
boto3==1.33.6
botocore==1.33.6
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dateutil==2.8.2
s3transfer==0.8.2
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
